+++ Tree Tracker/Extensions/PHAsset.swift
@@ -2,15 +2,15 @@ import Foundation
 import Photos
 
 extension PHAsset {
-    var _coordinates: String {
+    func stringifyCoordinates(roundingPrecision: Int = 5) -> String {
         if let coordinates = location?.coordinate {
-            return _stringify(coordinates: coordinates)
+            return _stringify(coordinates: coordinates, roundingPrecision: roundingPrecision)
         } else {
             return ""
         }
     }
 
-    private func _stringify(coordinates: CLLocationCoordinate2D) -> String {
-        return "\(coordinates.latitude), \(coordinates.longitude)"
+    private func _stringify(coordinates: CLLocationCoordinate2D, roundingPrecision: Int) -> String {
+        return String(format: "%.\(roundingPrecision)f, %.\(roundingPrecision)f", coordinates.latitude, coordinates.longitude)
     }
 }